Update patch set 1
Patch Set 1:

so, os-brick is the only package requiring >= 0.30.1 and we have 0.30.0. According to changelog in eventlet, we can live without those two patches and it seems that os-brick is just updating >= to reflect u_c instead of actual minimal version. I'm reluctant to update eventlet at this point of the cycle (it has been a pain in the past), may we do just >= 0.30.0?

Patch-set: 1
Label: Workflow=0